test: 회원가입, 로그인 테스트 간단하게
